requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
pymysql==1.1.0
cryptography==41.0.7
